fix: kronfluence language model test
