+++ tests/explainers/wrappers/test_kronfluence.py
@@ -1,18 +1,10 @@
 import pytest
-import torch.nn as nn
-from datasets import load_dataset
+
 from kronfluence.arguments import (  # type: ignore
     FactorArguments,
     ScoreArguments,
 )
 from kronfluence.utils.dataset import DataLoaderKwargs  # type: ignore
-from torch.utils.data import DataLoader
-from transformers import (
-    AutoConfig,
-    AutoModelForSequenceClassification,
-    AutoTokenizer,
-    DataCollatorWithPadding,
-)
 
 from quanda.explainers.wrappers import (
     Kronfluence,
@@ -26,7 +18,7 @@ from quanda.explainers.wrappers import (
     "test_id, model, dataset, test_tensor, test_labels, task",
     [
         (
-            "mnist",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "load_mnist_test_samples_1",
@@ -72,7 +64,7 @@ def test_kronfluence_explain(
     "test_id, model, dataset, task",
     [
         (
-            "mnist",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "classification_task",
@@ -109,7 +101,7 @@ def test_kronfluence_self_influence(
     "test_id, model, dataset, test_tensor, test_labels, task",
     [
         (
-            "mnist",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "load_mnist_test_samples_1",
@@ -154,7 +146,7 @@ def test_kronfluence_explain_functional(
     "test_id, model, dataset, task",
     [
         (
-            "mnist",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "classification_task",
@@ -190,7 +182,7 @@ def test_kronfluence_self_influence_functional(
     "test_id, model, dataset, test_tensor, test_labels, task, factor_args, score_args, dataloader_kwargs",
     [
         (
-            "mnist_optional",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "load_mnist_test_samples_1",
@@ -249,7 +241,7 @@ def test_kronfluence_explain_with_optional_args(
     "test_id, model, dataset, task, factor_args, score_args, dataloader_kwargs",
     [
         (
-            "mnist_optional",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "classification_task",
@@ -298,7 +290,7 @@ def test_kronfluence_self_influence_with_optional_args(
     "test_id, model, dataset, test_tensor, test_labels, task, factor_args, score_args",
     [
         (
-            "mnist_optional_functional",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "load_mnist_test_samples_1",
@@ -349,7 +341,7 @@ def test_kronfluence_explain_functional_with_optional_args(
     "test_id, model, dataset, task, factor_args, score_args",
     [
         (
-            "mnist_optional_functional",
+            "mnist_kronfluence",
             "load_mnist_model",
             "load_mnist_dataset",
             "classification_task",
@@ -386,159 +378,22 @@ def test_kronfluence_self_influence_functional_with_optional_args(
     ), "Self-influence scores have incorrect shape"
 
 
-# Partially copied from https://github.com/MadryLab/trak/blob/main/examples/qnli.py
-GLUE_TASK_TO_KEYS = {
-    "cola": ("sentence", None),
-    "mnli": ("premise", "hypothesis"),
-    "mrpc": ("sentence1", "sentence2"),
-    "qnli": ("question", "sentence"),
-    "qqp": ("question1", "question2"),
-    "rte": ("sentence1", "sentence2"),
-    "sst2": ("sentence", None),
-    "stsb": ("sentence1", "sentence2"),
-    "wnli": ("sentence1", "sentence2"),
-}
-
-TRAIN_SET_SIZE = 10
-VAL_SET_SIZE = 4
-
-
-class SequenceClassificationModel(nn.Module):
-    """
-    Wrapper for HuggingFace sequence classification models.
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.config = AutoConfig.from_pretrained(
-            "gchhablani/bert-base-cased-finetuned-qnli",
-            num_labels=2,
-            finetuning_task="qnli",
-            cache_dir=None,
-            revision="main",
-            token=None,
-        )
-
-        self.model = AutoModelForSequenceClassification.from_pretrained(
-            "gchhablani/bert-base-cased-finetuned-qnli",
-            config=self.config,
-            cache_dir=None,
-            revision="main",
-            token=None,
-            ignore_mismatched_sizes=False,
-        )
-
-        self.model.eval()
-
-    def forward(self, input_ids, token_type_ids, attention_mask):
-        return self.model(
-            input_ids=input_ids,
-            token_type_ids=token_type_ids,
-            attention_mask=attention_mask,
-        )
-
-
-def get_dataset(split, inds=None):
-    raw_datasets = load_dataset(
-        "glue",
-        "qnli",
-        cache_dir=None,
-        use_auth_token=None,
-    )
-    sentence1_key, sentence2_key = GLUE_TASK_TO_KEYS["qnli"]
-
-    tokenizer = AutoTokenizer.from_pretrained(
-        "gchhablani/bert-base-cased-finetuned-qnli",
-        cache_dir=None,
-        use_fast=True,
-        revision="main",
-        token=None,
-    )
-
-    padding = "max_length"
-    max_seq_length = 128
-
-    def preprocess_function(examples):
-        args = (
-            (examples[sentence1_key],)
-            if sentence2_key is None
-            else (examples[sentence1_key], examples[sentence2_key])
-        )
-        result = tokenizer(
-            *args, padding=padding, max_length=max_seq_length, truncation=True
-        )
-
-        result["labels"] = examples["label"]
-
-        return result
-
-    raw_datasets = raw_datasets.map(
-        preprocess_function,
-        batched=True,
-        load_from_cache_file=(not False),
-        desc="Running tokenizer on dataset",
-    )
-
-    if split == "train":
-        train_dataset = raw_datasets["train"]
-        ds = train_dataset
-    else:
-        eval_dataset = raw_datasets["validation"]
-        ds = eval_dataset
-    return ds
-
-
-def init_model(ckpt_path, device="cpu"):
-    model = SequenceClassificationModel()
-    return model
-
-
-def init_loaders(batch_size=2):
-    ds_train = get_dataset("train")
-    ds_train = ds_train.select(range(TRAIN_SET_SIZE))
-    ds_val = get_dataset("validation")
-    ds_val = ds_val.select(range(VAL_SET_SIZE))
-
-    tokenizer = AutoTokenizer.from_pretrained(
-        "gchhablani/bert-base-cased-finetuned-qnli",
-        cache_dir=None,
-        use_fast=True,
-        revision="main",
-        token=None,
-    )
-
-    data_collator = DataCollatorWithPadding(tokenizer=tokenizer)
-
-    return DataLoader(
-        ds_train,
-        batch_size=batch_size,
-        shuffle=False,
-        collate_fn=data_collator,
-    ), DataLoader(
-        ds_val, batch_size=batch_size, shuffle=False, collate_fn=data_collator
-    )
-
-
-def process_batch(batch):
-    return (
-        batch["input_ids"],
-        batch["token_type_ids"],
-        batch["attention_mask"],
-        batch["labels"],
-    )
-
-
 @pytest.mark.explainers
+@pytest.mark.parametrize(
+    "test_id, model, dataset",
+    [
+        (
+            "qnli_kronfluence",
+            "qnli_model",
+            "qnli_dataset",
+        ),
+    ],
+)
 def test_kronfluence_self_influence_qnli(
-    load_qnli_model,
-    text_classification_task,
+    test_id, model, dataset, text_classification_task, request
 ):
-    loader_train, loader_val = init_loaders()
-
-    model = init_model(".", "cpu")
-
-    train_dataset = loader_train.dataset
-    test_dataset = loader_val.dataset
+    model = request.getfixturevalue(model)
+    train_dataset, test_dataset = request.getfixturevalue(dataset)
 
     train_dataset.set_format(
         type="torch",
